Add a primitive particle writer.
This is implemented as a class method in Particles as
Particles::FormattedTableOutput().  Currently the data is dumped one
line per particle without headers and special formatting.  Each line
consists of id, x1, x2, x3, v1, v2, and v3 in order.

Whenever a formatted table output (tab) is requested, this class method
is called.