Overhaul particle data structure.
To be flexible for the derived classes to reuse the class and instance
methods, the particle data are now two 2D arrays, one for integers and
the other for reals.  The number of properties to be allocated is
initialized once when the class method Particles::Initialzed() is
called, where property index is asked for by calling
Particles::AddIntProperty() or Particles::AddRealProperty().

Any derived class can thus ask for more properties by simply overloading
Particles::Initialized() without changing any other methods for
communication purposes.

The particle IDs, positions, and velocities are still accessible by the
AthenaArrays pid, xp1, xp2, xp3, vp1, vp2, and vp3, which are shallow
copies of the 2D data arrays.