Particles outside meshblock & boundary conditions.
Now use the position indices to compute (again) the mesh coordinates of
a particle, when it is outside of the meshblock.

The class method Particles::ApplyBoundaryConditions() does not need to
check both sides of the domain.  Also, it is now returning a bool to
indicate if a particle is outside of the mesh boundary.  This is so that
one does not need to convert the particle coordinates back if it is
inside the boundary.