Change particle time step.
It was too expensive to consider acceleration.  The consideration will
be delegated to the derived class and the user who have a better handle
in considering acceleration-limited time step.

This change also leads to the removal of
Particles::InitialBlockTimeStep() from commit c53e1d3b, which is not
needed anymore.